+++ resources/examples/manager/BasicAssetLibrary/plugin/BasicAssetLibrary/bal.py
@@ -30,10 +30,6 @@ from collections import namedtuple
 from urllib.parse import urlparse
 
 
-from openassetio import constants
-from openassetio.exceptions import InvalidEntityReference
-
-
 EntityInfo = namedtuple("EntityInfo", ("name"), defaults=("",))
 Entity = namedtuple("Entity", ("traits"), defaults=({},))
 
@@ -80,8 +76,8 @@ def parse_entity_ref(entity_ref: str) -> EntityInfo:
     """
     uri_parts = urlparse(entity_ref)
 
-    if not uri_parts.path:
-        raise InvalidEntityReference("Missing entity name in path component", entity_ref)
+    if len(uri_parts.path) <= 1:
+        raise InvalidBALReference("Missing entity name in path component")
 
     # path will start with a /
     name = uri_parts.path[1:]
@@ -128,4 +124,8 @@ def managementPolicy(trait_set: set[str], library: dict) -> dict:
 
 
 class UnknownBALEntity(Exception):
+    pass
+
+
+class InvalidBALReference(Exception):
     pass